requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
av==16.1.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blis==1.3.3
build==1.4.0
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
chromadb==1.5.1
click==8.3.1
cloudpathlib==0.23.0
comm==0.2.3
confection==0.1.5
config==0.5.1
contourpy==1.3.3
cryptography==46.0.5
ctranslate2==4.7.1
cycler==0.12.1
cymem==2.0.13
dataclasses-json==0.6.7
datasets==4.6.0
debugpy==1.8.20
decorator==5.2.1
deep-translator==1.11.4
deprecated==1.3.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docstring-parser==0.17.0
durationpy==0.10
elastic-transport==9.2.1
elasticsearch==9.3.0
emoji==2.15.0
en-core-web-sm==3.8.0
et-xmlfile==2.0.0
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.133.0
faster-whisper==1.2.1
filelock==3.25.0
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-vision==3.12.1
googleapis-common-protos==1.72.0
groq==0.37.1
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hf-xet==1.3.2
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.5.0
idna==3.11
importlib-metadata==8.7.1
importlib-resources==6.5.2
installer==0.7.0
instructor==1.14.5
ipykernel==7.2.0
ipython==9.10.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.15
langchain-groq==1.1.2
langchain-huggingface==1.2.0
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langdetect==1.0.9
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.6
llvmlite==0.46.0
lxml==6.0.2
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml-dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
msoffcrypto-tool==6.0.0
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
oauthlib==3.3.1
olefile==0.47
ollama==0.6.1
onnx==1.20.1
onnxruntime==1.24.2
openai==2.24.0
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pdf2image==1.17.0
pdfminer-six==20221105
pexpect==4.9.0
pi-heif==1.3.0
pikepdf==10.3.0
pillow==12.1.1
platformdirs==4.9.2
posthog==5.4.0
preshed==3.0.12
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1-modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pygments==2.19.2
pymupdf==1.27.1
pypandoc-binary==1.16.2
pyparsing==3.3.2
pypdf==6.7.5
pypdfium2==5.6.0
pypika==0.51.1
pyproject-hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
python-pptx==1.0.2
pyyaml==6.0.3
pyzmq==27.1.0
ragas==0.4.3
rapidfuzz==3.14.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.5.1
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.47
srsly==2.5.2
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tesseract==0.1.3
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.25
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
unstructured==0.21.5
unstructured-client==0.42.10
unstructured-inference==1.5.2
unstructured-pytesseract==0.3.15
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.41.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.6.0
weasel==0.4.3
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
